refactor: 스웨거에서 static Request, Response 를 제대로 인식하지 못해서 분리
