+++ product/src/main/java/com/emotionalcart/adminproduct/presentation/AdminProductController.java
@@ -1,9 +1,11 @@
 package com.emotionalcart.adminproduct.presentation;
 
 import com.emotionalcart.adminproduct.application.AdminProductService;
-import com.emotionalcart.adminproduct.presentation.dto.CreateProduct;
+import com.emotionalcart.adminproduct.presentation.dto.CreateProductRequest;
+import com.emotionalcart.adminproduct.presentation.dto.CreateProductResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,9 +16,9 @@ public class AdminProductController implements AdminProductControllerDocs {
     private final AdminProductService adminProductService;
 
     @Override
-    @PostMapping
-    public ResponseEntity<CreateProduct.Response> createProduct(
-            @Valid @ModelAttribute CreateProduct.Request request) {
+    @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    public ResponseEntity<CreateProductResponse> createProduct(
+            @Valid @ModelAttribute CreateProductRequest request) {
         return ResponseEntity.ok(adminProductService.createProduct(request));
     }
 }
