+++ product/src/main/java/com/emotionalcart/adminproduct/presentation/dto/CreateProductResponse.java
@@ -1,0 +1,10 @@
+package com.emotionalcart.adminproduct.presentation.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CreateProductResponse {
+    private Long productId;
+}
