+++ product/src/main/java/com/emotionalcart/product/presentation/dto/response/CreateProductReviewResponse.java
@@ -1,0 +1,10 @@
+package com.emotionalcart.product.presentation.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CreateProductReviewResponse {
+    Long reviewId;
+}
